Fix timer ticking logic
previously, the timer would return the tick value iff there was a new tick
and otherwise -1. However, what I actually want is for tick() to reutrn
whatever the current tick value is if that value hasn't been returned before.
This has the effect of including the 0th tick in the return of tick().